+++ packages/react/List/src/components/Item/item.spec.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import { render, screen } from "@testing-library/react";
+
+import { Item } from "./Item";
+
+const content = "myText";
+
+describe("GIVEN <Item />", () => {
+  it("WHEN rendered THEN it should correctly render children elements", () => {
+    render(<Item>{content}</Item>);
+    expect(screen.getByText(content)).toBeTruthy();
+  });
+});