+++ packages/react/List/src/components/SkeletonItem/index.ts
@@ -1,0 +1,4 @@
+import { SkeletonItem } from "./SkeletonItem";
+
+export { SkeletonItem } from "./SkeletonItem";
+export default SkeletonItem;