+++ packages/styles/src/index.ts
@@ -13,4 +13,5 @@ export { tag } from "./packages/tag";
 export { text } from "./packages/text";
 export { title } from "./packages/title";
 export { toast } from "./packages/toast";
+export { list } from "./packages/list";
 export { toggle } from "./packages/toggle";